relative id is a long

+++ src/main/java/com/rootscity/model/PersonSummary.java
@@ -4,11 +4,11 @@ package com.rootscity.model;
  * Created by Nano on 06/01/2017.
  */
 public class PersonSummary {
-		public int id;
-		public NameForm nameParts;
-		public String gender;
-		public String birthDate;
-		public String birthPlace;
-		public String deathDate;
-		public String deathPlace;
+	public Long id;
+	public NameForm nameParts;
+	public String gender;
+	public String birthDate;
+	public String birthPlace;
+	public String deathDate;
+	public String deathPlace;
 }
